0.3.0 release

+++ entrypoint.sh
@@ -28,13 +28,32 @@ if [[ $AWS_SYNC_ENABLED == 'true' ]]; then
     
     if [[ $EFS_ENABLED == 'true' ]]; then
 
-        mkdir /cardano/db/$HOSTNAME/
-        cp -R /cardano/db/source/* /cardano/db/$HOSTNAME/
-        sed -i "s^/cardano/db^/cardano/db/$HOSTNAME^g" /cardano/scripts/.env 
+        if [[ $TESTNET_ENABLED == 'true' ]]; then
+
+            mkdir /cardano/db/$HOSTNAME/
+            cp -R /cardano/db/testnet/* /cardano/db/$HOSTNAME/
+            sed -i "s^/cardano/db^/cardano/db/$HOSTNAME^g" /cardano/scripts/.env 
+
+            echo "export TESTNET_ENABLED=true" >> /root/.bash_profile
+            echo "export TESTNET_ENABLED=true" >> /home/admin/.bash_profile
+
+        else
+
+            mkdir /cardano/db/$HOSTNAME/
+            cp -R /cardano/db/source/* /cardano/db/$HOSTNAME/
+            sed -i "s^/cardano/db^/cardano/db/$HOSTNAME^g" /cardano/scripts/.env 
+
+        fi
 
     else
 
-        aws s3 sync s3://$DB_BUCKET_NAME/ /cardano/db/
+        if [[ ! $REMOTE_URL_SYNC == 'true' ]]; then
+            if [[ $TESTNET_ENABLED == 'true' ]]; then
+                aws s3 sync s3://$DB_BUCKET_NAME/testnet/ /cardano/db/
+            else
+                aws s3 sync s3://$DB_BUCKET_NAME/ /cardano/db/
+            fi
+        fi
     
     fi
 
@@ -43,12 +62,29 @@ if [[ $AWS_SYNC_ENABLED == 'true' ]]; then
 
     if [[ $MASTER_NODE == 'true' ]]; then
         if [[ $EFS_ENABLED == 'true' ]]; then
-            echo "0 0 * * * source /root/.bash_profile && aws s3 sync /cardano/db/$HOSTNAME/ s3://$DB_BUCKET_NAME/ --delete &>>/var/log/cron.log" >> /var/spool/cron/root
-            echo "0 15 * * * source /root/.bash_profile && rm -rf /cardano/db/source/ && mkdir /cardano/db/source/ && cp -R /cardano/db/$HOSTNAME/* /cardano/db/source/ &>>/var/log/cron.log" >> /var/spool/cron/root
+            if [[ $TESTNET_ENABLED == 'true' ]]; then
+                echo "0 0 * * * source /root/.bash_profile && aws s3 sync /cardano/db/$HOSTNAME/ s3://$DB_BUCKET_NAME/testnet/ --delete &>>/var/log/cron.log" >> /var/spool/cron/root
+                echo "0 15 * * * source /root/.bash_profile && rm -rf /cardano/db/testnet/ && mkdir /cardano/db/testnet/ && cp -R /cardano/db/$HOSTNAME/* /cardano/db/testnet/ &>>/var/log/cron.log" >> /var/spool/cron/root
+            else
+                echo "0 0 * * * source /root/.bash_profile && aws s3 sync /cardano/db/$HOSTNAME/ s3://$DB_BUCKET_NAME/ --delete &>>/var/log/cron.log" >> /var/spool/cron/root
+                echo "0 15 * * * source /root/.bash_profile && rm -rf /cardano/db/source/ && mkdir /cardano/db/source/ && cp -R /cardano/db/$HOSTNAME/* /cardano/db/source/ &>>/var/log/cron.log" >> /var/spool/cron/root
+            fi
         else
-            echo "0 0 * * * source /root/.bash_profile && aws s3 sync /cardano/db/ s3://$DB_BUCKET_NAME/ --delete &>>/var/log/cron.log" >> /var/spool/cron/root
+            if [[ $TESTNET_ENABLED == 'true' ]]; then
+                echo "0 0 * * * source /root/.bash_profile && aws s3 sync /cardano/db/ s3://$DB_BUCKET_NAME/testnet/ --delete &>>/var/log/cron.log" >> /var/spool/cron/root
+            else
+                echo "0 0 * * * source /root/.bash_profile && aws s3 sync /cardano/db/ s3://$DB_BUCKET_NAME/ --delete &>>/var/log/cron.log" >> /var/spool/cron/root
+            fi
         fi
     fi
+fi
+
+if [[ $REMOTE_URL_SYNC == 'true' ]]; then
+
+    curl -L -o ./db_archive.tar.gz $REMOTE_DB_URL
+    tar -xvf ./db_archive.tar.gz --directory ${NODE_HOME}/db/
+    rm -rf ./db_archive.tar.gz
+
 fi
 
 nohup crond >>/var/log/cron.log 2>&1 &
@@ -58,12 +94,16 @@ nohup cardano-submit-api --mainnet --socket-path $CARDANO_NODE_SOCKET_PATH --con
 
 if [[ $DB_SYNC_ENABLED == 'true' ]]; then
 
-    if [[ $RESTORE_DB_SYNC_SNAPSHOT == 'true' ]]; then
+    if [[ $MASTER_NODE == 'true' ]]; then
 
-        echo -e "\n-= Download most recent cardano-db-sync snapshot"
-        sudo curl -L -o cardano-snapshot.tgz https://update-cardano-mainnet.iohk.io/cardano-db-sync/12/db-sync-snapshot-schema-12-block-6793499-x86_64.tgz
-        tar -xvf cardano-snapshot.tgz --directory /cardano/snapshots --exclude configuration
-        rm -rf cardano-snapshot.tgz
+        if [[ $RESTORE_DB_SYNC_SNAPSHOT == 'true' ]]; then
+
+            echo -e "\n-= Download most recent cardano-db-sync snapshot"
+            curl -L -o cardano-snapshot.tgz https://update-cardano-mainnet.iohk.io/cardano-db-sync/12/db-sync-snapshot-schema-12-block-6850499-x86_64.tgz
+            tar -xvf cardano-snapshot.tgz --directory /cardano/snapshots --exclude configuration
+            rm -rf cardano-snapshot.tgz
+
+        fi
 
     fi
 
@@ -74,7 +114,14 @@ if [[ $DB_SYNC_ENABLED == 'true' ]]; then
     sed -i "s^username^${POSTGRES_USER}^g" /cardano/config/pgpass-mainnet
     sed -i "s^password^${POSTGRES_PASS}^g" /cardano/config/pgpass-mainnet
 
-    nohup bash -c '/cardano/scripts/start-db-sync.sh' >>/var/log/dbsync.log 2>&1 &
+    if [[ $MASTER_NODE == 'true' ]]; then
+
+        nohup bash -c '/cardano/scripts/start-db-sync.sh' >>/var/log/dbsync.log 2>&1 &
+
+    fi
+
+    echo "export DB_SYNC_ENABLED=true" >> /root/.bash_profile
+    echo "export DB_SYNC_ENABLED=true" >> /home/admin/.bash_profile
 
 fi
